refactor(common): add docker-compose dev environment
