+++ logs/bem-server_out.log
@@ -1,0 +1,2 @@
+listening on *:4000
+Database Connected!
