+++ monitor/html/config.js
