+++ monitor/html/js/check_nodes.js
