+++ monitor/html/js/utils.js
