+++ monitor/html/style.css
